@@ -8,7 +8,10 @@
     "build-action-upload-android": "ncc build src/upload-android/main.ts -m -o lib/upload-android",
     "build-action-upload-ios": "ncc build src/upload-ios/main.ts -m -o lib/upload-ios",
     "compile": "npm-run-all --parallel build-action-build-ios build-action-upload-android build-action-upload-ios",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "lint": "npx eslint src/.",
+    "format:write": "npx prettier --write src/.",
+    "format:check": "npx prettier --check src/."
   },
   "keywords": [],
   "author": "",
@@ -23,16 +26,18 @@
     "p-timeout": "^6.1.2"
   },
   "devDependencies": {
-    "npm-run-all": "4.1.5",
+    "@eslint/js": "^9.17.0",
     "@types/jest": "^29.5.12",
     "@types/node": "^20.11.17",
     "@typescript-eslint/eslint-plugin": "^6.21.0",
     "@typescript-eslint/parser": "^6.21.0",
     "@vercel/ncc": "^0.38.1",
-    "eslint": "^8.56.0",
+    "eslint": "^8.57.1",
     "github-action-ts-run-api": "^3.0.4",
     "jest": "^29.7.0",
+    "npm-run-all": "4.1.5",
     "ts-jest": "^29.1.2",
-    "typescript": "^5.3.3"
+    "typescript": "^5.7.2",
+    "typescript-eslint": "^8.18.1"
   }
 }
