@@ -1,5 +1,5 @@
-const core = require("@actions/core");
-const exec = require("@actions/exec");
+import * as core from "@actions/core";
+import * as exec from "@actions/exec";
 
 async function run() {
   try {
@@ -24,9 +24,14 @@ async function run() {
     // Validate increment build number
     if (
       core.getInput("increment-build-number") === "testflight" &&
-      (!core.getInput("bundle-identifier") || !core.getInput("app-store-connect-api-key-id") || !core.getInput("app-store-connect-api-key-issuer-id") || !core.getInput("app-store-connect-api-key-base64"))
+      (!core.getInput("bundle-identifier") ||
+        !core.getInput("app-store-connect-api-key-id") ||
+        !core.getInput("app-store-connect-api-key-issuer-id") ||
+        !core.getInput("app-store-connect-api-key-base64"))
     ) {
-      throw new Error("increment-build-number='testflight' requires 'bundle-identifier', 'app-store-connect-api-key-id', 'app-store-connect-api-key-issuer-id' and 'app-store-connect-api-key-base64' to be provided.");
+      throw new Error(
+        "increment-build-number='testflight' requires 'bundle-identifier', 'app-store-connect-api-key-id', 'app-store-connect-api-key-issuer-id' and 'app-store-connect-api-key-base64' to be provided.",
+      );
     }
 
     // Set environment variables
@@ -34,7 +39,7 @@ async function run() {
     process.env.P12_KEY_BASE64 = core.getInput("p12-key-base64");
     process.env.P12_CER_BASE64 = core.getInput("p12-cer-base64");
     process.env.MOBILEPROVISION_BASE64 = core.getInput(
-      "mobileprovision-base64"
+      "mobileprovision-base64",
     );
     process.env.P12_PATH = core.getInput("p12-path");
     process.env.P12_KEY_PATH = core.getInput("p12-key-path");
@@ -53,27 +58,37 @@ async function run() {
     process.env.DISABLE_TARGETS = core.getInput("disable-targets");
     process.env.EXPORT_OPTIONS = core.getInput("export-options");
     process.env.CLONED_SOURCE_PACKAGES_PATH = core.getInput(
-      "cloned-source-packages-path"
+      "cloned-source-packages-path",
     );
     process.env.BUILD_SDK = core.getInput("build-sdk");
     process.env.BUILD_DESTINATION = core.getInput("build-destination");
     process.env.ENTITLMENTS_FILE_PATH = core.getInput("entitlements-file-path");
-    process.env.INCREMENT_BUILD_NUMBER = core.getInput("increment-build-number");
-    process.env.INCREMENT_VERSION_NUMBER = core.getInput('increment-version-number');
-    process.env.BUNDLE_IDENTIFIER = core.getInput('bundle-identifier');
-    process.env.APP_STORE_CONNECT_API_KEY_ID = core.getInput('app-store-connect-api-key-id');
-    process.env.APP_STORE_CONNECT_API_KEY_ISSUER_ID = core.getInput('app-store-connect-api-key-issuer-id');
-    process.env.APP_STORE_CONNECT_API_KEY_BASE64 = core.getInput('app-store-connect-api-key-base64');
-    process.env.BUILD_PATH = core.getInput('build-path');
-    process.env.CUSTOM_KEYCHAIN_NAME = core.getInput('custom-keychain-name');
+    process.env.INCREMENT_BUILD_NUMBER = core.getInput(
+      "increment-build-number",
+    );
+    process.env.INCREMENT_VERSION_NUMBER = core.getInput(
+      "increment-version-number",
+    );
+    process.env.BUNDLE_IDENTIFIER = core.getInput("bundle-identifier");
+    process.env.APP_STORE_CONNECT_API_KEY_ID = core.getInput(
+      "app-store-connect-api-key-id",
+    );
+    process.env.APP_STORE_CONNECT_API_KEY_ISSUER_ID = core.getInput(
+      "app-store-connect-api-key-issuer-id",
+    );
+    process.env.APP_STORE_CONNECT_API_KEY_BASE64 = core.getInput(
+      "app-store-connect-api-key-base64",
+    );
+    process.env.BUILD_PATH = core.getInput("build-path");
+    process.env.CUSTOM_KEYCHAIN_NAME = core.getInput("custom-keychain-name");
 
     // Execute build.sh
     await exec.exec(`bash ${__dirname}/../../src/build-ios/build.sh`);
   } catch (error: unknown) {
     if (error instanceof Error) {
-        core.setFailed(error.message)
+      core.setFailed(error.message);
     } else {
-        core.setFailed('Unknown error occurred.')
+      core.setFailed("Unknown error occurred.");
     }
   }
 }
